Keep static mutation cleanup failures visible
